pyproject.toml
[tool.poetry]
name = "scorecard"
version = "0.1.0"
description = ""
authors = ["Vernon Swanepoel <frantic.naturalist@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.3.5"
matplotlib = "^3.5.1"
jupyter = "^1.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
